+++ src/mako/application/cli/commands/app/GenerateSecret.php
@@ -9,6 +9,7 @@ namespace mako\application\cli\commands\app;
 
 use mako\application\Application;
 use mako\file\FileSystem;
+use mako\reactor\attributes\Command as CommandAttribute;
 use mako\reactor\Command;
 use mako\security\Key;
 
@@ -17,13 +18,9 @@ use function preg_replace;
 /**
  * Command that generates a new application secret.
  */
+#[CommandAttribute('app:generate-secret', 'Generates a new application secret.')]
 class GenerateSecret extends Command
 {
-	/**
-	 * {@inheritDoc}
-	 */
-	protected string $description = 'Generates a new application secret.';
-
 	/**
 	 * Executes the command.
 	 *