+++ src/mako/application/cli/commands/app/ListRoutes.php
@@ -14,6 +14,8 @@ use mako\http\routing\Dispatcher;
 use mako\http\routing\Route;
 use mako\http\routing\Router;
 use mako\http\routing\Routes;
+use mako\reactor\attributes\Arguments;
+use mako\reactor\attributes\Command as CommandAttribute;
 use mako\reactor\Command;
 use mako\utility\Arr;
 
@@ -28,24 +30,13 @@ use function str_pad;
 /**
  * Command that lists all registered routes.
  */
+#[CommandAttribute('app:routes', 'Lists all registered routes.')]
+#[Arguments(
+	new Argument('-f|--filter', 'Filter routes using the route action, name or path', Argument::IS_OPTIONAL),
+	new Argument('-d|--detailed', 'Show more information about the route', Argument::IS_BOOL | Argument::IS_OPTIONAL),
+)]
 class ListRoutes extends Command
 {
-	/**
-	 * {@inheritDoc}
-	 */
-	protected string $description = 'Lists all registered routes.';
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function getArguments(): array
-	{
-		return [
-			new Argument('-f|--filter', 'Filter routes using the route action, name or path', Argument::IS_OPTIONAL),
-			new Argument('-d|--detailed', 'Show more information about the route', Argument::IS_BOOL | Argument::IS_OPTIONAL),
-		];
-	}
-
 	/**
 	 * Returns a normalzed action name.
 	 */