+++ src/mako/application/cli/commands/cache/Clear.php
@@ -9,27 +9,18 @@ namespace mako\application\cli\commands\cache;
 
 use mako\cache\CacheManager;
 use mako\cli\input\arguments\Argument;
+use mako\reactor\attributes\Arguments;
+use mako\reactor\attributes\Command as CommandAttribute;
 
 /**
  * Command that clears the cache.
  */
+#[CommandAttribute('cache:clear', 'Clears the cache.')]
+#[Arguments(
+	new Argument('-c|--configuration', 'Configuration name', Argument::IS_OPTIONAL),
+)]
 class Clear extends Command
 {
-	/**
-	 * {@inheritDoc}
-	 */
-	protected string $description = 'Clears the cache.';
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function getArguments(): array
-	{
-		return [
-			new Argument('-c|--configuration', 'Configuration name', Argument::IS_OPTIONAL),
-		];
-	}
-
 	/**
 	 * Executes the command.
 	 *