+++ src/mako/application/cli/commands/migrations/Reset.php
@@ -9,29 +9,20 @@ namespace mako\application\cli\commands\migrations;
 
 use mako\application\cli\commands\migrations\traits\RollbackTrait;
 use mako\cli\input\arguments\Argument;
+use mako\reactor\attributes\Arguments;
+use mako\reactor\attributes\Command as CommandAttribute;
 
 /**
  * Command that rolls back the last batch of migrations.
  */
+#[CommandAttribute('migration:reset', 'Resets the database schema.')]
+#[Arguments(
+	new Argument('-d|--database', 'Sets which database connection to use', Argument::IS_OPTIONAL),
+	new Argument('-f|--force', 'Force the schema reset?', Argument::IS_BOOL),
+)]
 class Reset extends Command
 {
 	use RollbackTrait;
-
-	/**
-	 * {@inheritDoc}
-	 */
-	protected string $description = 'Resets the database schema.';
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function getArguments(): array
-	{
-		return [
-			new Argument('-d|--database', 'Sets which database connection to use', Argument::IS_OPTIONAL),
-			new Argument('-f|--force', 'Force the schema reset?', Argument::IS_BOOL),
-		];
-	}
 
 	/**
 	 * Executes the command.