+++ src/mako/application/cli/commands/migrations/Status.php
@@ -8,6 +8,8 @@
 namespace mako\application\cli\commands\migrations;
 
 use mako\cli\input\arguments\Argument;
+use mako\reactor\attributes\Arguments;
+use mako\reactor\attributes\Command as CommandAttribute;
 
 use function count;
 use function vsprintf;
@@ -15,24 +17,13 @@ use function vsprintf;
 /**
  * Command that checks if there are any outstanding migrations.
  */
+#[CommandAttribute('migration:status', 'Checks if there are any outstanding migrations.')]
+#[Arguments(
+	new Argument('-d|--database', 'Sets which database connection to use', Argument::IS_OPTIONAL),
+	new Argument('-e|--exit-code', 'Exits with 1 if there are outstanding migrations and 0 if there are none', Argument::IS_BOOL),
+)]
 class Status extends Command
 {
-	/**
-	 * {@inheritDoc}
-	 */
-	protected string $description = 'Checks if there are any outstanding migrations.';
-
-	/**
-	 * {@inheritDoc}
-	 */
-	public function getArguments(): array
-	{
-		return [
-			new Argument('-d|--database', 'Sets which database connection to use', Argument::IS_OPTIONAL),
-			new Argument('-e|--exit-code', 'Exits with 1 if there are outstanding migrations and 0 if there are none', Argument::IS_BOOL),
-		];
-	}
-
 	/**
 	 * Executes the command.
 	 */