+++ CHANGELOG.md
@@ -43,6 +43,8 @@ The major version bump is due to upping the required PHP version from `8.1` to `
 * Added `FileSystem::clearCache()` method.
 * Added `FileInfo::hasPermissions()` method that accepts an integer or a `Permissions` instance.
 * Added `FileInfo::getPermissions()` method that returns a `Permissions` instance.
+* Reactor commands names and descriptions can now be registered with the `Command` attribute.
+* Reactor command arguments can now be registered with the `Arguments` attribute.
 
 #### Changes
 
@@ -53,11 +55,21 @@ The major version bump is due to upping the required PHP version from `8.1` to `
 * Renamed the `ErrorHandler::handler()` method to `ErrorHandler::handle()`.
 * The
 
+#### Deprecations
+
+* Deprecated the `Command::$command` property. Use the `Command` attribute instead.
+* Deprecated the `Command::$description` property. Use the `Command` attribute instead.
+* Deprecated the `Command::getCommand()` method.
+* Deprecated the `Command::getDescription()` method.
+* Deprecated the `Command::getArguments()` method. Use the `Arguments` attribute instead.
+
 #### Improvements
 
 * Various improvements and optimizations.
 
 > Check out the upgrade guide for details on how to upgrade from `10.0.*.`
+
+> All deprecated features will be removed in Mako 12.0.
 
 --------------------------------------------------------